% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet_utils.R
\name{build_normals}
\alias{build_normals}
\title{Build Annual Averages and Period Normals}
\usage{
build_normals(var, years = c(2001:2020), dir = geo_path())
}
\arguments{
\item{var}{what variables to download (1-many)}

\item{dir}{the directory to cache data. Default is \code{geo_path()}}

\item{year}{what years (1979-2020)}
}
\value{
file to disk
}
\description{
Calculate the annual mean from raw GridMet data for the years requested. Afterward, compute a normal for that period.
}
\seealso{
Other gridmet: 
\code{\link{calculate_ai}()},
\code{\link{downloadGridMET}()}
}
\concept{gridmet}
