% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet_utils.R
\name{calculate_ai}
\alias{calculate_ai}
\title{Generate Aridity Index (PET/PR)}
\usage{
calculate_ai(years = c(2001:2020), dir = geo_path())
}
\arguments{
\item{years}{what years (1979-2020). \code{pr} and \code{pet} data must already be downloaded. See \code{downloadGridMET()}.}

\item{dir}{the directory to cache data. Default is \code{geo_path()}}
}
\description{
Generate Aridity Index (PET/PR)
}
\seealso{
Other gridmet: 
\code{\link{build_normals}()},
\code{\link{downloadGridMET}()}
}
\concept{gridmet}
