% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{day8_to_month}
\alias{day8_to_month}
\title{8day to monthly mean}
\usage{
day8_to_month(
  dir = geo_path(),
  product = NULL,
  date = NULL,
  prefix = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dir}{the directory to cache data. Default is \code{geo_path()}}

\item{product}{the MODIS product to search}

\item{date}{the date of the months to process to build passed as a vector of length 1 (single date),
or 2 (range of dates)}

\item{prefix}{the prefix to search for}

\item{overwrite}{should the files be overwriten?}
}
\description{
takes 8 day MODIS tiles and generates a monthly mean
mean((tile / 8)) / days_in_month
}
\seealso{
Other modis: 
\code{\link{downloadMODIS}()},
\code{\link{mosaicMODIS}()}
}
\concept{modis}
