% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet_utils.R
\name{downloadGridMET}
\alias{downloadGridMET}
\title{Download GridMet Daily files}
\usage{
downloadGridMET(var, year, dir = geo_path())
}
\arguments{
\item{var}{what variables to download (1-many)}

\item{year}{what years (1979-2020)}

\item{dir}{the directory to cache data. Default is \code{geo_path()}}
}
\value{
file to disk
}
\description{
Daily gridmet files have daily values, and each year is stored as
a seperate file.
}
\seealso{
Other gridmet: 
\code{\link{build_normals}()},
\code{\link{calculate_ai}()}
}
\concept{gridmet}
