% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{downloadMODIS}
\alias{downloadMODIS}
\title{Download MODIS tiles}
\usage{
downloadMODIS(
  AOI,
  product,
  date = NULL,
  dir = geo_path(),
  netrc = getNetrcPath()
)
}
\arguments{
\item{AOI}{the region to find data for (sf)}

\item{product}{a MODIS product}

\item{date}{the date(s) to download data for passed as a vector of length 1 (single date), or
2 (range of dates)}

\item{dir}{the directory to cache data. Default is \code{geo_path()}}

\item{netrc}{the path to your netrc file. Defaults to \code{getNetrcPath()}}
}
\description{
Download MODIS tiles
}
\seealso{
Other modis: 
\code{\link{day8_to_month}()},
\code{\link{mosaicMODIS}()}
}
\concept{modis}
