% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{geo_cache_list}
\alias{geo_cache_list}
\title{Manage cached files}
\usage{
geo_cache_list(pattern = NULL)
}
\arguments{
\item{pattern}{pattern to look for using list.files}
}
\description{
Manage cached files
}
