% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_abstraction.R
\name{geogrid_warp}
\alias{geogrid_warp}
\title{Wrap}
\usage{
geogrid_warp(file, grid = NULL, r = "near", filename = NULL)
}
\arguments{
\item{file}{a vector of files paths}

\item{grid}{a grid abstraction to warp to (see \code{make_grid()})}

\item{r}{resampling method see gdalwarp for options}

\item{filename}{character. Output filename}
}
\value{
SpatRast object
}
\description{
Wrap
}
\seealso{
Other grid: 
\code{\link{getSubsets}()},
\code{\link{make_grid}()}
}
\concept{grid}
