% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_utils.R
\name{get_ned}
\alias{get_ned}
\title{Download the 1 (~30 meter) or 1/3 (~10 meter) arc-second National Elevation Dataset.}
\usage{
get_ned(AOI, res = "1", outdir = "/Volumes/Transcend/ngen/DEM")
}
\arguments{
\item{AOI}{An \code{sf} object to serve as a template for cropping.}

\item{res}{A character string representing the desired resolution of the NED. '1'
indicates the 1 arc-second NED (the default), while '13' indicates the 1/3 arc-second dataset.}

\item{outdir}{A character string indicating where raw downloaded files should be put.}
}
\description{
Download the 1 (~30 meter) or 1/3 (~10 meter) arc-second National Elevation Dataset.
}
\concept{Elevation}
