% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{getSubsets}
\alias{getSubsets}
\title{Find HDF subsets}
\usage{
getSubsets(file = NULL, product = NULL)
}
\arguments{
\item{file}{a HDF file path}

\item{product}{a MODIS product name}
}
\description{
Find HDF subsets
}
\seealso{
Other grid: 
\code{\link{geogrid_warp}()},
\code{\link{make_grid}()}
}
\concept{grid}
