% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet_utils.R
\name{jennings_snow_threshold}
\alias{jennings_snow_threshold}
\title{Jennings Snow Threshold}
\usage{
jennings_snow_threshold(dir = geo_path())
}
\arguments{
\item{dir}{the directory to cache data. Default is \code{geo_path()}}
}
\value{
file to Jennings Snow Threshold
}
\description{
\href{https://doi.org/10.5061/dryad.c9h35}{Jennings, Keith S.; Winchell, Taylor S.; Livneh, Ben; Molotch, Noah P. (2019), Data from: Spatial variation of the rain-snow temperature threshold across the Northern Hemisphere, Dryad, Dataset, }
}
\concept{snow}
