% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_abstraction.R
\name{make_grid}
\alias{make_grid}
\title{Build Grid Abstraction}
\usage{
make_grid(file = NULL, ext = NULL, dimXY = NULL, resXY = NULL, prj = NULL)
}
\arguments{
\item{file}{file path}

\item{ext}{c(xmin, ymin, xmax, ymax)}

\item{dimXY}{c(width, height)}

\item{resXY}{c(xres, yres)}

\item{prj}{srs_def}
}
\value{
list
}
\description{
Define a grid abstraction for a file or an extent, res/dim, and projection
}
\seealso{
Other grid: 
\code{\link{geogrid_warp}()},
\code{\link{getSubsets}()}
}
\concept{grid}
