% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{mosaicMODIS}
\alias{mosaicMODIS}
\title{Mosaic/Warp Raw MODIS Tiles}
\usage{
mosaicMODIS(
  dir = geo_path(),
  product,
  date = NULL,
  pattern = NULL,
  grid = NULL,
  r = "near",
  overwrite = FALSE,
  scale_factor = 1,
  range = NULL,
  cog = TRUE
)
}
\arguments{
\item{dir}{the directory to cache data. Default is \code{geo_path()}}

\item{product}{the MODIS product to search for}

\item{date}{the date of the mosaic to build passed as a vector of length 1 (single date), or
2 (range of dates)}

\item{pattern}{the HDF file pattern to extract (see \code{getSubsets()})}

\item{grid}{a grid abstraction build with \code{make_grid()}}

\item{r}{GDAL resampling method}

\item{overwrite}{should the files be overwriten?}

\item{scale_factor}{product scale factor}

\item{range}{valid data range}

\item{cog}{should a COG directory be built?}
}
\description{
Mosaics all MODIS HDF files for a given date to a new file with a flexible extent (te),
resolution (tr), CRS (t_srs). If GDAL options are left NULL, the GDAL defaults are used.
}
\seealso{
Other modis: 
\code{\link{.modis_dir}()},
\code{\link{day8_to_month}()},
\code{\link{downloadMODIS}()}
}
\concept{modis}
