% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{day8_to_month}
\alias{day8_to_month}
\title{8day to monthly mean}
\usage{
day8_to_month(
  dir = "/Volumes/Transcend/ngen/MODIS/MOD15A2H.006/mosaics_cogs",
  measurement = "representative",
  date = NULL,
  overwrite = FALSE,
  cog = TRUE
)
}
\arguments{
\item{dir}{the directory with 8 day data}

\item{measurement}{are the 8 day files a single best \code{representative} (default) or a daily \code{sum}}

\item{date}{the date of the months to process to build passed as a vector of length 1 (single date),
or 2 (range of dates)}

\item{overwrite}{should the files be overwriten?}

\item{cog}{should a COG file (Cloud Optimized Geotif) be produced?}
}
\description{
takes 8 day MODIS tiles and generates a monthly mean
mean((tile / 8)) / days_in_month
}
\seealso{
Other modis: 
\code{\link{.modis_dir}()},
\code{\link{downloadMODIS}()},
\code{\link{mosaicMODIS}()}
}
\concept{modis}
