% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{.modis_dir}
\alias{.modis_dir}
\title{Scrape MODIS molt and mota archives}
\usage{
.modis_dir(product)
}
\arguments{
\item{product}{a MODIS product}
}
\description{
Scrape MODIS molt and mota archives
}
