% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{downloadMODIS}
\alias{downloadMODIS}
\title{Download MODIS tiles}
\usage{
downloadMODIS(
  AOI,
  product,
  date = NULL,
  dir = geo_path(),
  netrc = getNetrcPath()
)
}
\arguments{
\item{AOI}{the region to find data for (sf)}

\item{product}{a MODIS product}

\item{dir}{the directory to cache data. Default is \code{geo_path()}}

\item{netrc}{the path to your netrc file. Defaults to \code{getNetrcPath()}}

\item{startDate}{a date to start looking for data}

\item{endDate}{a data to end looking for data. If NULL defaults to startDate}
}
\value{

}
\description{
Download MODIS tiles
}
