% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_abstraction.R
\name{geogrid_warp}
\alias{geogrid_warp}
\title{Wrap}
\usage{
geogrid_warp(files, grid = NULL, r = "near", disk = TRUE)
}
\arguments{
\item{files}{a vector of files paths}

\item{grid}{a grid abstraction to warp to (see \code{make_grid()})}

\item{r}{resampling method see gdalwarp for options}

\item{disk}{return file paths (TRUE) or raster stack (FALSE)}
}
\value{
vecotr of filepaths or raster::stack
}
\description{
Wrap
}
\seealso{
Other grid: 
\code{\link{getSubsets}()},
\code{\link{make_grid}()}
}
\concept{grid}
