% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet_utils.R
\name{jennings_snow_threshold}
\alias{jennings_snow_threshold}
\title{Jennings Snow Threshold}
\usage{
jennings_snow_threshold(dir = geo_path())
}
\arguments{
\item{dir}{the directory to cache data. Default is \code{geo_path()}}
}
\value{
file to Jennings Snow Threshold
}
\description{
Jennings Snow Threshold
}
\concept{snow}
