% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_utils.R
\name{mosaicMODIS}
\alias{mosaicMODIS}
\title{Mosaic/Warp Raw MODIS Tiles}
\usage{
mosaicMODIS(
  dir = geo_path(),
  prefix = "",
  product,
  date = NULL,
  pattern = NULL,
  te = NULL,
  tr = NULL,
  t_srs = NULL,
  r = NULL,
  overwrite = FALSE,
  of = "GTiff"
)
}
\arguments{
\item{dir}{the directory to cache data. Default is \code{geo_path()}}

\item{prefix}{the prefix for the outputfiles, shoudl be specfic to the requested grid}

\item{product}{the MODIS product to search for}

\item{date}{the date of the mosaic to build passed as a vector of length 1 (single date), or
2 (range of dates)}

\item{pattern}{the HDF file pattern to extract (see \code{getSubsets()})}

\item{te}{GDAL target extent. Vector of length 4 (xmin,ymin,xmax,ymax)}

\item{tr}{GDAL target extent. Vector of length 2 (xres, yres}

\item{t_srs}{GDAL target projection}

\item{r}{GDAL resampling method}

\item{overwrite}{should the files be overwriten?}

\item{of}{GDAL output file type (default = GTiff)}
}
\value{

}
\description{
Mosaics all MODIS HDF files for a given date to a new file with a flexible extent (te),
resolution (tr), CRS (t_srs). If GDAL options are left NULL, the GDAL defaults are used.
}
